test: move allowHTTP1 header tests to a focused http2 test file
